requirements.txt
apache-airflow>=2.6.0
requests>=2.28.0
pendulum>=2.0.0 